+++ src/builder/webpack/config.browser.js
@@ -2,7 +2,7 @@
 
 import { join } from "path"
 import commonWebpackConfig from "./config.common.js"
-import { offlinePlugin, offlineEntry } from "../../_utils/offline/webpack.js"
+import { offlinePlugin } from "../../_utils/offline/webpack.js"
 import PhenomicLoaderWebpackPlugin from "../../loader/plugin.js"
 
 const chunkNameBrowser = "phenomic.browser"
@@ -24,7 +24,6 @@ export default (config: PhenomicConfig): WebpackConfig => {
 
       [chunkNameBrowser]: [
         join(config.cwd, config.scriptBrowser),
-        ...offlineEntry(config),
       ],
     },
   }